Change by Codex: Extract shared report image thumbnail plumbing
- Replace user-specific adapter with shared read-only thumbnail adapter.
- Extract cache file naming into ReportImageCache helper (stable naming).
- Wire shared adapter into existing user/manager detail fragments without behavior change.

Prompt: You are working in the SeeSomethingABQ Android app module.

Goal:
Refine the shared attached-image display plumbing created for the user detail page so it is cleanly reusable by the manager detail page in the next prompt. This prompt should stay focused on shared display architecture, not new features.

Current background:
The user report detail page now downloads attached report images via the authenticated image endpoint, writes them to app cache, and displays thumbnails in a recycler.

Current relevant files:
- User detail fragment/layout
- Any new image adapter, view holder, UI-state model, helper, service, or viewmodel classes created in the prior prompt
- Existing manager detail fragment/layout:
  - src/main/java/edu/cnm/deepdive/seesomethingabq/controller/ManagerReportDetailFragment.kt
  - src/main/res/layout/fragment_manager_report_detail.xml

Requirements:
1. Review the newly added user-side image thumbnail code and refactor only where necessary so the display plumbing can be reused by both user and manager detail screens.

2. Prefer shared components for read-only image display such as:
   - thumbnail item layout
   - recycler adapter / view holder
   - lightweight UI-state model
   - image file/cache helper
   - authenticated image download service/helper

3. Keep user-specific editing concerns out of the shared read-only display layer.
   - This shared layer should support read-only thumbnail display only.
   - Future user-side add/remove/edit image flows can build on top later without forcing manager-specific code into the shared layer.

4. Do not change the user-visible behavior from the previous prompt unless needed for correctness or reuse.
   - User detail should still behave the same after this refactor.

5. Do not add manager image display in this prompt yet.
   - This prompt is only to make the shared code clean enough for the mirror pass.

Implementation guidance:
- Avoid broad refactors that create extra abstractions without payoff.
- Prefer small, concrete shared pieces over a complicated generic framework.
- Preserve the current cache directory/file naming approach.
- Preserve the current placeholder/failure behavior.

Validation target:
- App build passes.
- User detail image thumbnails still work exactly as before.
- Shared code is clearly reusable by the manager detail screen.

At the end:
- Summarize which shared pieces were introduced or cleaned up.
- Explicitly note what remains manager-specific vs shared.
